Add checkout automation for iPhone with shipping and payment flow

+++ Checkout Automation - iPhone/src/data/dataHandler.js
@@ -1,24 +1,22 @@
 const fs = require('fs');
 const path = require('path');
 
-const dataPath = path.join(__dirname, 'data.json');
+const dataPath = path.join(__dirname, '../../payment-data.json');
 
-function readData() {
+function savePaymentData(data) {
+  fs.writeFileSync(dataPath, JSON.stringify(data, null, 2));
+}
+
+function loadPaymentData() {
   try {
     return JSON.parse(fs.readFileSync(dataPath, 'utf8'));
-  } catch (err) {
+  } catch (error) {
     return null;
   }
 }
 
-function writeData(data) {
-  fs.writeFileSync(dataPath, JSON.stringify(data, null, 2));
+function clearPaymentData() {
+  if (fs.existsSync(dataPath)) fs.unlinkSync(dataPath);
 }
 
-function clearData() {
-  if (fs.existsSync(dataPath)) {
-    fs.unlinkSync(dataPath);
-  }
-}
-
-module.exports = { readData, writeData, clearData };
+module.exports = { savePaymentData, loadPaymentData, clearPaymentData };
